uploader/requirements.txt
github3.py==1.3.0
